+++ Job-Dam/Sources/Service/Request/Request.swift
@@ -2,5 +2,5 @@ import Foundation
 
 struct Evaluation: Codable {
     let id: Int
-    let point: Int
+    let point: Double
 }
